requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
aioice=0.7.6=pypi_0
aiortc=1.4.0=pypi_0
av=10.0.0=pypi_0
blinker=1.5=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py310h06a4308_0
cryptography=39.0.1=pypi_0
dnspython=2.3.0=pypi_0
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-crc32c=1.5.0=pypi_0
grpcio=1.51.1=pypi_0
idna=3.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.4.1=pypi_0
ncurses=6.4=h6a678d5_0
netifaces=0.11.0=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.68=pypi_0
openssl=1.1.1s=h7f8727e_0
pexpect=4.8.0=pypi_0
pip=23.0=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyee=9.0.4=pypi_0
pylibsrtp=0.8.0=pypi_0
pyopenssl=23.0.0=pypi_0
python=3.10.9=h7a1cb2a_0
pytz=2022.7.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
sentencepiece=0.1.97=pypi_0
setuptools=65.6.3=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.40.1=h5082296_0
streamlit=1.18.1=pypi_0
streamlit-option-menu=0.3.2=pypi_0
streamlit-webrtc=0.44.5=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizer=3.4.2=pypi_0
torch=1.13.1=pypi_0
tqdm=4.64.1=pypi_0
tzdata=2022g=h04d1e81_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h5eee18b_1
zlib=1.2.13=h5eee18b_0
